+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- method `IAVLTree.isEmpty()`
+
 ## [0.1.1] (2021-04-25)
 
 ### Fixed